webapi/tct-alarm-tizen-tests/alarm/AlarmAbsolute_constructor_date_daysOfTheWeek.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>
<head>
<title>AlarmAbsolute_constructor_date_daysOfTheWeek</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: AlarmAbsolute_constructor_date_daysOfTheWeek
//==== LABEL Test whether the constructor with date parameter and day of the week parameter are supported
//==== SPEC Tizen Web API:Application:Alarm:AlarmAbsolute:constructor C
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/alarm.html
//==== PRIORITY P1
//==== TEST_CRITERIA CONSTRA CONSTRM

test(function () {
    var i, alarm, date = new Date(),
        daysOfTheWeekPossibilities = ["MO", "TU", "WE", "TH", "FR", "SA", "SU"],
        daysOfTheWeek = [];
    date.setFullYear(date.getFullYear() + 1);
    date.setMilliseconds(0);
    for (i = 0; i < daysOfTheWeekPossibilities.length; i++) {
        daysOfTheWeek.push(daysOfTheWeekPossibilities[i]);
        alarm = new tizen.AlarmAbsolute(date, daysOfTheWeek);
        assert_true(alarm instanceof tizen.AlarmAbsolute, "object was not created properly");
        assert_equals(alarm.date.getTime(), date.getTime(), "date passed as constructor parameter is different than date in the object");
        assert_array_equals(alarm.daysOfTheWeek, daysOfTheWeek,
            "days of the week array passed as constructor parameter is different than days of the week array in the object");
    }
}, document.title);

</script>
</body>
</html>
